secret_santa_bot/pyproject.toml
[tool.poetry]
name = "secret-santa-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
psutil = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
